+++ lua/plugins/config/languages/init.lua
@@ -1,0 +1,3 @@
+return function(use)
+	require('plugins.config.languages.golang')(use)
+end